+++ database/migrations/2021_06_15_025547_create_newsletters_table.php
@@ -1,0 +1,17 @@
+<?php
+
+use Illuminate\Database\Migrations\Migration;
+use Illuminate\Database\Schema\Blueprint;
+use Illuminate\Support\Facades\Schema;
+
+class CreateNewslettersTable extends Migration
+{
+    public function up()
+    {
+        Schema::create('newsletters', function (Blueprint $table) {
+            $table->id();
+            $table->string('email');
+            $table->timestamps();
+        });
+    }
+}