+++ resources/views/components/bit/logo.blade.php
@@ -9,6 +9,6 @@
     @if($mobile)
     <span class="sr-only">Harriet's Library</span>
     @else
-    <span class="block pb-3 -mb-4 text-2xl font-bold text-transparent bg-clip-text bg-gradient-to-r from-teal-200 to-cyan-400 sm:pb-5">Harriet's Library</span>
+    <span class="block pb-3 -mb-4 text-2xl font-bold text-transparent bg-clip-text bg-gradient-to-r from-teal-400 to-cyan-400 dark:from-teal-200 dark:to-cyan-400 sm:pb-5">Harriet's Library</span>
     @endif
 </a>