+++ CHANGELOG.md
@@ -10,17 +10,24 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 - Drag to reorder
 - Move up/down queue
 - Parent/child accounts
-- Notifications
 
 ---
 
-## Unpublished
+## [0.1.1] - 2021-06-15
 
 ### Added
 
 - Minus button on book if it is in queue
 - Add search to filters on discover
 - View book details
+- Notifications
+- Add basic newsletter signup
+
+### Fixed
+
+- Mobile Nav
+- Mobile spacing for search box on discover page
+- Mobile spacing on book page
 
 ---